requirements.txt
PyYAML>=6.0
python-docx>=1.1.2
requests>=2.32.0
httpx>=0.27.0
beautifulsoup4>=4.12.3
playwright>=1.52.0
trafilatura>=1.10.0
lxml>=5.2.2
pypdf>=5.1.0
Pillow>=10.4.0
numpy>=1.26.0
pypdfium2>=4.30.0
paddleocr>=2.8.1
rapidocr-onnxruntime>=1.3.25
opencv-python>=4.10.0
chromadb>=0.5.5
sentence-transformers>=3.0.1
PyMuPDF>=1.24.0
Jinja2>=3.1.4
premailer>=3.10.0
